requirements.txt
attrs==21.4.0
black==21.12b0
click==8.0.3
greenlet==1.1.2
iniconfig==1.1.1
isort==5.10.1
kafka-python==2.0.2
msgpack==1.0.3
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
pluggy==1.0.0
py==1.11.0
pynvim==0.4.3
pyparsing==3.0.6
pytest==6.2.5
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
